requirements.txt
# GUI
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0

# System Monitoring
psutil==7.0.0

# Core Utilities
numpy==1.26.4
pandas==2.2.2
pillow==11.2.0
python-dateutil==2.9.0.post0

# Web Scraping
selenium==4.18.1
beautifulsoup4==4.12.3
webdriver-manager==4.0.1
lxml==5.2.0
requests==2.31.0

# Development Tools
colorama==0.4.6
python-dotenv==1.0.1
